requirement.txt
colorama==0.3.9
enum34==1.1.6
lxml==4.1.1
namedlist==1.7
py==1.4.34
pytest==2.9.0
pytest-allure-adaptor==1.7.9
selenium==3.7.0
six==1.11.0
